+++ packages/core/src/generate/custom.ts
@@ -1,4 +1,4 @@
-import type { ERC1155Options } from '../erc1155';
+import type { CustomOptions } from '../custom';
 import { accessOptions } from '../set-access-control';
 import { infoOptions } from '../set-info';
 import { upgradeableOptions } from '../set-upgradeable';
@@ -7,18 +7,13 @@ import { generateAlternatives } from './alternatives';
 const booleans = [true, false];
 
 const blueprint = {
-  name: ['MyToken'],
-  uri: ['https://example.com/'],
-  burnable: booleans,
+  name: ['MyContract'],
   pausable: booleans,
-  mintable: booleans,
-  supply: booleans,
-  updatableUri: booleans,
   access: accessOptions,
   upgradeable: upgradeableOptions,
   info: infoOptions,
 };
 
-export function* generateERC1155Options(): Generator<Required<ERC1155Options>> {
+export function* generateCustomOptions(): Generator<Required<CustomOptions>> {
   yield* generateAlternatives(blueprint);
 }