+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.1 (2022-06-30)
+
+- Support custom contract type, optional access control. ([#112](https://github.com/OpenZeppelin/contracts-wizard/pull/112))
+
 ## 0.1.0 (2022-06-15)
 
 - Initial API for Solidity. ([#136](https://github.com/OpenZeppelin/contracts-wizard/pull/136))